pyproject.toml
[project]
name = "Advent_Of_Code"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "requests",
    "python-dotenv",
    "matplotlib",
]